requirements.txt
certifi==2020.12.5
elasticsearch==7.10.1
gitdb==4.0.5
gitdb2==4.0.2
GitPython==3.0.6
Jinja2==2.11.1
MarkupSafe==1.1.1
PyYAML==5.3.1
smmap==3.0.4
tqdm==4.54.1
truffleHog==2.1.11
truffleHogRegexes==0.0.7
urllib3==1.26.2
